cambios a la ventana principal

+++ backend/ventana_principal_backend.py
@@ -1,10 +1,11 @@
 from threading import Thread
+from os import path
 
 from PyQt5.QtCore import QObject, pyqtSignal
 
 from utils.find_key import find_key
 from utils.find_length import find_length
-from utils.vigenere import decode
+from utils.vigenere import decode, most_probable_length
 
 
 class VentanaPrincipalBackend(QObject):
@@ -13,41 +14,61 @@ class VentanaPrincipalBackend(QObject):
     senal_set_text = pyqtSignal(str)
     senal_set_recomended = pyqtSignal(str)
 
+    senal_boton_archivo = pyqtSignal(bool)
+    senal_boton_encuentra_largo = pyqtSignal(bool)
+    senal_boton_encuentra_key = pyqtSignal(bool)
+    senal_boton_decodificar = pyqtSignal(bool)
+
     def __init__(self):
         super().__init__()
         self.texto_codificado = ''
         self.clave_encontrada = ''
         self.largos_posibles = []
+        self.file_path = ''
 
     def set_file(self, path):
+        self.senal_boton_decodificar.emit(False)
+
         self.senal_add_to_table.emit([(0, 0) for _ in range(4)])
         self.senal_set_text.emit('')
         self.texto_codificado = ''
         self.clave_encontrada = ''
         self.largos_posibles = []
+        self.file_path = ''
 
         with open(path) as file:
+            self.file_path = path
             self.texto_codificado = file.read()
+            self.texto_codificado = ''.join(
+                [letter.upper() for letter in self.texto_codificado if letter.isalpha()])
+
+        self.senal_boton_decodificar.emit(True)
 
     def start_find_length(self):
         thread = Thread(target=self.find_length)
         thread.start()
 
     def find_length(self):
+        self.senal_boton_encuentra_largo.emit(False)
+
         if self.texto_codificado:
             self.largos_posibles = find_length(self.texto_codificado)
             self.senal_add_to_table.emit(self.largos_posibles)
-            
-            self.senal_set_recomended.emit(self.len_factible())
+
+            self.senal_set_recomended.emit(self.most_probable_length())
 
         else:
             self.senal_pop_up.emit('Se requiere un un texto para decodificar')
+
+        self.senal_boton_encuentra_largo.emit(True)
 
     def start_find_key(self, largo):
         thread = Thread(target=self.find_key, args=(largo,))
         thread.start()
 
     def find_key(self, largo):
+        self.senal_boton_encuentra_key.emit(False)
+
         if self.texto_codificado and largo > 0:
             self.clave_encontrada = find_key(
                 self.texto_codificado, largo_clave=largo, lang='ENG')
@@ -62,19 +83,25 @@ class VentanaPrincipalBackend(QObject):
                     'Necesitas introducir o generar una clave')
                 self.senal_set_text.emit('')
 
+        self.senal_boton_encuentra_key.emit(True)
+
     def decodificar(self):
-        if self.clave_encontrada:
-            texto_decodificado = decode(
-                self.texto_codificado, self.clave_encontrada)
+        self.senal_boton_decodificar.emit(False)
+
+        if self.clave_encontrada and path.exists(self.file_path):
+            with open(self.file_path) as f:
+                texto = f.read()
+
+            texto_decodificado, _ = decode(
+                texto, self.clave_encontrada)
             with open('decoded_text.txt', 'w') as f:
                 f.write(texto_decodificado)
         else:
-            self.senal_pop_up.emit('Debes completar pasos previos')
-    
-    def len_factible(self):
-        largos_factibles = []
-        for index in range(len(self.largos_posibles)):
-            largos_factibles.append(self.largos_posibles[index])
-            if not self.largos_posibles[index][1] < self.largos_posibles[index + 1][1] * 18:
-                break
-        return str(max(largos_factibles)[0])
+            if not path.exists(self.file_path):
+                self.senal_pop_up.emit('ARCHIVO DEJO DE EXISTIR')
+            else:
+                self.senal_pop_up.emit('Debes completar pasos previos')
+        self.senal_boton_decodificar.emit(True)
+
+    def most_probable_length(self):
+        return str(most_probable_length(self.largos_posibles))
